change codes for switch profile

+++ includes/modules/api/del_profile.php
@@ -10,5 +10,9 @@ $query = $DB->prepare($sql);
 $query->execute();
 $user = $query->fetch(PDO::FETCH_ASSOC);
 
-if ($user['user_id'] != $profile_id)
+$used_in_profile_id = $user['user_id'];
+if (isset($infos->used_in_profile))
+    $used_in_profile_id = $infos->used_in_profile;
+
+if ($user['user_id'] != $profile_id && $used_in_profile_id != $profile_id)
     deleteUserFinaly($profile_id);
